+++ tests/simulcast.spec.ts
@@ -3,7 +3,7 @@ import { expect, test } from '@playwright/test'
 test('simulcast sendonly/recvonly pages', async ({ page }) => {
   await page.goto('http://localhost:9000/simulcast/')
 
-  await page.click('#start')
+  await page.click('#connect')
 
   // 安全によせて 5 秒待つ
   await page.waitForTimeout(5000)
@@ -74,5 +74,5 @@ test('simulcast sendonly/recvonly pages', async ({ page }) => {
   expect(sendonlyVideoR2OutboundRtpStats?.packetsSent).toBeGreaterThan(0)
   expect(sendonlyVideoR2OutboundRtpStats?.scalabilityMode).toEqual('L1T1')
 
-  await page.click('#stop')
+  await page.click('#disconnect')
 })